+++ event.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/bjartek/underflow"
 	"github.com/onflow/flow-go-sdk"
 	"github.com/sanity-io/litter"
 	"golang.org/x/exp/slices"
@@ -22,7 +23,6 @@ type OverflowEventList []OverflowEvent
 type OverflowEvents map[string]OverflowEventList
 
 func (me OverflowEvents) GetStakeholders(stakeholders map[string][]string) map[string][]string {
-
 	for _, events := range me {
 		for _, event := range events {
 			eventStakeholders := event.GetStakeholders()
@@ -121,11 +121,11 @@ func parseEvents(events []flow.Event, idPrefix string) (OverflowEvents, Overflow
 		for id, field := range event.Value.Fields {
 			name := fieldNames[id]
 
-			adr := ExtractAddresses(field)
+			adr := underflow.ExtractAddresses(field)
 			if len(adr) > 0 {
 				addresses[name] = adr
 			}
-			value := CadenceValueToInterface(field)
+			value := underflow.CadenceValueToInterface(field)
 			if value != nil {
 				finalFields[name] = value
 			}
@@ -196,7 +196,6 @@ var feeReceipients = []string{"0xf919ee77447b7497", "0x912d5440f7e3769e", "0xe5a
 
 // Filtter out fee events
 func (overflowEvents OverflowEvents) FilterFees(fee float64, payer string) OverflowEvents {
-
 	filteredEvents := overflowEvents
 	for name, events := range overflowEvents {
 		if strings.HasSuffix(name, "FlowFees.FeesDeducted") {
@@ -255,6 +254,7 @@ func printOrLog(t *testing.T, s string) {
 		t.Helper()
 	}
 }
+
 func (overflowEvents OverflowEvents) Print(t *testing.T) {
 	if t != nil {
 		t.Helper()
@@ -285,7 +285,7 @@ func (overflowEvents OverflowEvents) FilterEvents(ignoreFields OverflowEventFilt
 	filteredEvents := OverflowEvents{}
 	for name, events := range overflowEvents {
 
-		//find if we should ignore fields
+		// find if we should ignore fields
 		ignoreFieldNames := []string{}
 		for ignoreEvent, fields := range ignoreFields {
 			if strings.HasSuffix(name, ignoreEvent) {
