+++ script.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"testing"
 
+	"github.com/bjartek/underflow"
 	"github.com/enescakir/emoji"
 	"github.com/fatih/color"
 	"github.com/hexops/autogold"
@@ -31,7 +32,6 @@ type OverflowScriptOptsFunction func(opts ...OverflowInteractionOption) *Overflo
 
 // compose interactionOptions into a new Script function
 func (o *OverflowState) ScriptFN(outerOpts ...OverflowInteractionOption) OverflowScriptFunction {
-
 	return func(filename string, opts ...OverflowInteractionOption) *OverflowScriptResult {
 		outerOpts = append(outerOpts, opts...)
 		return o.Script(filename, outerOpts...)
@@ -40,9 +40,7 @@ func (o *OverflowState) ScriptFN(outerOpts ...OverflowInteractionOption) Overflo
 
 // compose fileName and interactionOptions into a new Script function
 func (o *OverflowState) ScriptFileNameFN(filename string, outerOpts ...OverflowInteractionOption) OverflowScriptOptsFunction {
-
 	return func(opts ...OverflowInteractionOption) *OverflowScriptResult {
-
 		outerOpts = append(outerOpts, opts...)
 		return o.Script(filename, outerOpts...)
 	}
@@ -62,11 +60,9 @@ func (o *OverflowState) Script(filename string, opts ...OverflowInteractionOptio
 		panic(result.Err)
 	}
 	return result
-
 }
 
 func (fbi *OverflowInteractionBuilder) runScript() *OverflowScriptResult {
-
 	o := fbi.Overflow
 	osc := &OverflowScriptResult{Input: fbi}
 	if fbi.Error != nil {
@@ -107,7 +103,7 @@ func (fbi *OverflowInteractionBuilder) runScript() *OverflowScriptResult {
 	}
 	result, err := o.Flowkit.ExecuteScript(fbi.Ctx, script, *sc)
 	osc.Result = result
-	osc.Output = CadenceValueToInterface(result)
+	osc.Output = underflow.CadenceValueToInterface(result)
 	if err != nil {
 		osc.Err = errors.Wrapf(err, "scriptFileName:%s", fbi.FileName)
 	}
@@ -147,7 +143,6 @@ type OverflowScriptResult struct {
 }
 
 func (osr *OverflowScriptResult) PrintArguments(t *testing.T) {
-
 	args := osr.Input.NamedCadenceArguments
 	maxLength := 0
 	for name := range args {
@@ -159,7 +154,7 @@ func (osr *OverflowScriptResult) PrintArguments(t *testing.T) {
 	format := fmt.Sprintf("%%%ds -> %%v", maxLength)
 
 	for name, arg := range args {
-		value, err := CadenceValueToJsonString(arg)
+		value, err := underflow.CadenceValueToJsonString(arg)
 		if err != nil {
 			panic(err)
 		}
@@ -173,7 +168,6 @@ func (osr *OverflowScriptResult) GetAsJson() (string, error) {
 		return "", errors.Wrapf(osr.Err, "script: %s", osr.Input.FileName)
 	}
 	j, err := json.MarshalIndent(osr.Output, "", "    ")
-
 	if err != nil {
 		return "", errors.Wrapf(err, "script: %s", osr.Input.FileName)
 	}
@@ -286,7 +280,6 @@ func (osr *OverflowScriptResult) MarshalPointerAs(pointer string, marshalTo inte
 
 // get the given jsonPointer as interface{}
 func (osr *OverflowScriptResult) GetWithPointer(pointer string) (interface{}, error) {
-
 	ptr, err := gojsonpointer.NewJsonPointer(pointer)
 	if err != nil {
 		return nil, err
